requirements.txt
# Core MCP dependencies
mcp==1.0.0
pydantic>=2.5.0
httpx>=0.25.0
python-dotenv>=1.0.0

# API client generation
openapi-python-client>=0.15.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Logging and monitoring
structlog>=23.1.0

# Additional utilities
pyyaml>=6.0.1
click>=8.1.0
rich>=13.0.0
